Added TODOs

+++ src/database/db.py
@@ -1,4 +1,4 @@
-from sqlalchemy import create_engine, Column, Integer, String, Text, DateTime
+from sqlalchemy import create_engine, Column, Integer, String, Text, DateTime, Float
 from sqlalchemy.orm import sessionmaker, declarative_base
 
 from datetime import datetime
@@ -20,7 +20,7 @@ class PageModel(Base):
     id = Column(Integer, primary_key=True)
 
     status_code = Column(Integer)
-    elapsed = Column(Integer)
+    elapsed = Column(Float)
     crawled_at = Column(DateTime, default=datetime.utcnow)
 
     url = Column(String)
